fix compile warning and uninitialized member variable ActionTimeline::_loop

+++ extensions/cocostudio/ActionTimeline/CCActionTimeline.h
@@ -206,7 +206,7 @@ protected:
     cocos2d::Vector<Timeline*> _timelineList;
 
     int     _duration;
-    double  _time;
+    float   _time;
     float   _timeSpeed;
     float   _frameInternal;
     bool    _playing;
